Rename reference from welcome4521 to welcome and correct activity layout.
