+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.0"
     defaultConfig {
-        applicationId "com.example.welcome4521"
+        applicationId "com.example.hkust.comp4521.welcome"
         minSdkVersion 15
         targetSdkVersion 29
         versionCode 1
