+++ app/src/main/java/com/example/welcome/welcome.java
@@ -1,10 +1,11 @@
-package com.example.welcome4521;
+package com.example.welcome;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class MainActivity extends AppCompatActivity {
+
+public class welcome extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
